cleanup(record): lift `Tristate` to `types`
It's not uncommon that the caller would want to know whether all items are selected/unselected, so we expose this functionality in `types`.
